… src/client/Microsoft.Identity.Client/Utils/EnumerableExtensions.cs