Fixes #2716 

**Changes proposed in this request**
Replace IEnumerable with IReadOnly list to avoid re-execution of filters 
Add more logging
Short-circuit some logic if token_count = 0

**Testing**
perf testing

**Performance impact**
FindAccessTokenAsync for client_creds flow (before and after) does not show any meaningful difference.

before
|                     Method | TokenCacheSize |         Mean |        Error |       StdDev |
|--------------------------- |--------------- |-------------:|-------------:|-------------:|
| 'Different tenants - O(1)' |           1000 |     73.24 us |     1.372 us |     1.468 us |
|  'Different scopes - O(n)' |           1000 | 58,334.41 us | 1,104.858 us | 2,647.169 us |

after
|                     Method | TokenCacheSize |         Mean |        Error |       StdDev |
|--------------------------- |--------------- |-------------:|-------------:|-------------:|
| 'Different tenants - O(1)' |           1000 |     76.41 us |     0.513 us |     0.429 us |
|  'Different scopes - O(n)' |           1000 | 58,034.85 us | 1,149.229 us | 1,951.479 us |